requirements.txt
streamlit>=1.35.0
google-generativeai>=0.7.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pdfplumber>=0.10.0
pypdf>=4.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
plotly>=5.18.0
openpyxl>=3.1.0
schedule>=1.2.0
psycopg2-binary>=2.9.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
